requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==12.0
python-dotenv==1.0.1
anthropic>=0.39.0
httpx==0.27.2
